@@ -8,13 +8,12 @@ import javax.annotation.Nullable;
 import io.github.tofodroid.mods.mimi.common.tile.AConfigurableTile;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -32,7 +31,7 @@ public abstract class AConfigurableTileBlock<B extends AConfigurableTile> extend
     protected abstract void appendSettingsTooltip(ItemStack blockItemStack, List<Component> tooltip);
 
     @Override
-    public InteractionResult use(BlockState state, Level worldIn, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit) {
+    public InteractionResult useWithoutItem(BlockState state, Level worldIn, BlockPos pos, Player player, BlockHitResult hit) {
         B tile = getTileForBlock(worldIn, pos);
         
         if(tile != null) {
@@ -46,8 +45,8 @@ public abstract class AConfigurableTileBlock<B extends AConfigurableTile> extend
     }
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable BlockGetter blockGetter, List<Component> tooltip, TooltipFlag flag) {
-        super.appendHoverText(stack, blockGetter, tooltip, flag);
+    public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltip, TooltipFlag flag) {
+        super.appendHoverText(stack, context, tooltip, flag);
         this.appendSettingsTooltip(stack, tooltip);
     }
     
@@ -56,7 +55,7 @@ public abstract class AConfigurableTileBlock<B extends AConfigurableTile> extend
         B tileEntity = getTileForBlock(worldIn, pos);
         if (tileEntity instanceof B) {
             ItemStack newStack = new ItemStack(stack.getItem(), 1);
-            newStack.setTag(stack.getOrCreateTag().copy());
+            newStack.applyComponents(stack.getComponents());
             ((B)tileEntity).setSourceStack(newStack);
         }
     }
