@@ -4,7 +4,6 @@ import io.github.tofodroid.mods.mimi.common.container.ContainerTuningTable;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.SimpleMenuProvider;
@@ -24,7 +23,7 @@ public class BlockTuningTable extends Block {
    }
 
    @Override
-   public InteractionResult use(BlockState state, Level worldIn, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit) {
+   public InteractionResult useWithoutItem(BlockState state, Level worldIn, BlockPos pos, Player player, BlockHitResult hit) {
       if (worldIn.isClientSide) {
          return InteractionResult.SUCCESS;
       } else {
