@@ -2,8 +2,6 @@ package io.github.tofodroid.mods.mimi.common.config.instrument;
 
 import java.util.List;
 
-import io.github.tofodroid.mods.mimi.common.item.IColorableItem;
-
 public class InstrumentSpec {
     public Byte instrumentId;
     public Integer midiBankNumber;
@@ -19,6 +17,6 @@ public class InstrumentSpec {
     }
 
     public Integer defaultColor() {
-        return isColorable() ? (defaultColor != null ? defaultColor : IColorableItem.DEFAULT_WHITE_COLOR) : null;
+        return isColorable() ? (defaultColor != null ? defaultColor : -1) : null;
     }
 }
