@@ -8,6 +8,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
+import net.minecraft.network.syncher.SynchedEntityData.Builder;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.Level;
@@ -87,7 +88,7 @@ public class EntityNoteResponsiveTile extends Entity {
     }
 
     @Override
-    protected void defineSynchedData() { }
+    protected void defineSynchedData(Builder pBuilder) { }
 
     @Override
     protected void readAdditionalSaveData(CompoundTag tag) {
