@@ -10,6 +10,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
+import net.minecraft.network.syncher.SynchedEntityData.Builder;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.player.Player;
@@ -78,10 +79,10 @@ public class EntitySeat extends Entity {
         }
     }
 
-    @Override
-    public float ridingOffset(Entity e) {
-        return -0.2f;
-    }
+    // @Override
+    // public float ridingOffset(Entity e) {
+    //     return -0.2f;
+    // }
     
     public BlockPos getSource() {
         return source;
@@ -145,5 +146,5 @@ public class EntitySeat extends Entity {
     }
 
     @Override
-    protected void defineSynchedData() { }
+    protected void defineSynchedData(Builder pBuilder) { }
 }
