@@ -1,7 +1,8 @@
 package io.github.tofodroid.mods.mimi.common.item;
 
-public interface IInstrumentItem extends IColorableItem {
+public interface IInstrumentItem {
     public Byte getInstrumentId();
     public Integer getDefaultChannels();
     public String getRegistryName();
+    public Integer getDefaultColor();
 }
