@@ -1,7 +1,5 @@
 package io.github.tofodroid.mods.mimi.common.item;
 
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.ListTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
@@ -15,13 +13,11 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 
 import java.util.List;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 
 import io.github.tofodroid.mods.mimi.client.gui.ClientGuiWrapper;
 import io.github.tofodroid.mods.mimi.common.config.ConfigProxy;
@@ -36,7 +32,6 @@ public class ItemInstrumentHandheld extends Item implements IInstrumentItem {
     protected final Byte instrumentId;
     protected final Integer defaultChannels;
     protected final Integer defaultColor;
-    protected final Boolean colorable;
 
     public ItemInstrumentHandheld(Properties props, Byte instrumentId) {
         super(props.stacksTo(1));
@@ -44,8 +39,7 @@ public class ItemInstrumentHandheld extends Item implements IInstrumentItem {
         this.REGISTRY_NAME = spec.registryName;
         this.defaultChannels = MidiNbtDataUtils.getDefaultChannelsForBank(spec.midiBankNumber);
         this.instrumentId = instrumentId;
-        this.colorable = spec.isColorable();
-        this.defaultColor = this.colorable ? spec.defaultColor() : null;
+        this.defaultColor = spec.defaultColor();
     }
 
     public ItemInstrumentHandheld(Properties props, InstrumentSpec spec) {
@@ -53,23 +47,13 @@ public class ItemInstrumentHandheld extends Item implements IInstrumentItem {
         this.REGISTRY_NAME = spec.registryName;
         this.defaultChannels = MidiNbtDataUtils.getDefaultChannelsForBank(spec.midiBankNumber);
         this.instrumentId = spec.instrumentId;
-        this.colorable = spec.isColorable();
-        this.defaultColor = this.colorable ? spec.defaultColor() : null;
+        this.defaultColor = spec.defaultColor();
     }
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {
-        super.appendHoverText(stack, worldIn, tooltip, flagIn);
-
-        // Client-side only
-        if(worldIn != null && worldIn.isClientSide) {
-            MidiNbtDataUtils.appendSettingsTooltip(stack, tooltip);
-        }
-    }
-
-    @Override
-    public Boolean isColorable() {
-        return this.colorable;
+    public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltip, TooltipFlag flagIn) {
+        super.appendHoverText(stack, context, tooltip, flagIn);
+        MidiNbtDataUtils.appendSettingsTooltip(stack, tooltip);
     }
 
     @Override
@@ -82,13 +66,13 @@ public class ItemInstrumentHandheld extends Item implements IInstrumentItem {
         return this.instrumentId;
     }
 
-    @Override
-    public InteractionResult useOn(UseOnContext context) {
-        if(washItem(context)) {
-            return InteractionResult.SUCCESS;
-        }
-        return super.useOn(context);
-    }
+    // @Override
+    // public InteractionResult useOn(UseOnContext context) {
+    //     if(washItem(context)) {
+    //         return InteractionResult.SUCCESS;
+    //     }
+    //     return super.useOn(context);
+    // }
     
     @Override
     public EquipmentSlot getEquipmentSlot(ItemStack stack) {
@@ -127,27 +111,6 @@ public class ItemInstrumentHandheld extends Item implements IInstrumentItem {
         }
 
 		return new InteractionResultHolder<>(InteractionResult.PASS, playerIn.getItemInHand(handIn));
-    }
-
-    @Override
-    public void verifyTagAfterLoad(CompoundTag tag) {
-        if(tag == null) return;
-
-        if(tag.contains("inventory")) {
-            ListTag listtag = tag.getCompound("inventory").getList("Items", 10);
-
-            if(listtag != null && listtag.size() > 0) {
-                for(int i = 0; i < listtag.size(); ++i) {
-                    CompoundTag stackTag = listtag.getCompound(i);
-                    String itemId = stackTag.getString("id");
-
-                    if(itemId.equalsIgnoreCase("mimi:switchboard") && stackTag.contains("tag", 10)) {
-                        tag = tag.merge(MidiNbtDataUtils.convertSwitchboardToDataTag(stackTag.getCompound("tag")));
-                        tag.remove("inventory");
-                    }
-                }
-            }
-        }
     }
 
     public static ItemStack getEntityHeldInstrumentStack(LivingEntity entity, InteractionHand handIn) {
