@@ -14,6 +14,7 @@ import net.minecraft.resources.ResourceLocation;
 
 public class MultiMidiNotePacket implements CustomPacketPayload {
     public static final ResourceLocation ID = new ResourceLocation(MIMIMod.MODID, MultiMidiNotePacket.class.getSimpleName().toLowerCase());
+    public static final CustomPacketPayload.Type<MultiMidiNotePacket> TYPE = new Type<>(ID);
 
     private final Map<Long, ArrayList<NetMidiEvent>> sourceMap;
     public final TreeMap<Long, List<MidiNotePacket>> resultPackets;
@@ -36,15 +37,10 @@ public class MultiMidiNotePacket implements CustomPacketPayload {
         this.resultPackets = packets;
         this.sourceMap = Map.of();
     }
-
+    
     @Override
-    public ResourceLocation id() {
-        return MultiMidiNotePacket.ID;
-    }
-
-    @Override
-    public void write(FriendlyByteBuf buf) {
-        MultiMidiNotePacket.encodePacket(this, buf);
+    public Type<? extends CustomPacketPayload> type() {
+       return TYPE;
     }
 
     public static MultiMidiNotePacket decodePacket(FriendlyByteBuf buf) {
