@@ -7,6 +7,7 @@ import net.minecraft.resources.ResourceLocation;
 
 public class ServerTimeSyncPacket implements CustomPacketPayload {
     public static final ResourceLocation ID = new ResourceLocation(MIMIMod.MODID, ServerTimeSyncPacket.class.getSimpleName().toLowerCase());
+    public static final CustomPacketPayload.Type<ServerTimeSyncPacket> TYPE = new Type<>(ID);
     public final long currentServerMilli;
     public final Boolean firstRequest;
 
@@ -18,15 +19,10 @@ public class ServerTimeSyncPacket implements CustomPacketPayload {
         this.currentServerMilli = currentServerMilli;
         this.firstRequest = firstRequest;
     }
-
+    
     @Override
-    public ResourceLocation id() {
-        return ServerTimeSyncPacket.ID;
-    }
-
-    @Override
-    public void write(FriendlyByteBuf buf) {
-        ServerTimeSyncPacket.encodePacket(this, buf);
+    public Type<? extends CustomPacketPayload> type() {
+       return TYPE;
     }
 
     public static ServerTimeSyncPacket decodePacket(FriendlyByteBuf buf) {
