@@ -5,9 +5,9 @@ import java.util.List;
 
 import io.github.tofodroid.mods.mimi.common.block.AColoredBlock;
 import io.github.tofodroid.mods.mimi.util.TagUtils;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.core.NonNullList;
-import net.minecraft.core.RegistryAccess;
-import net.minecraft.nbt.CompoundTag;
+import net.minecraft.core.component.DataComponentType;
 import net.minecraft.world.inventory.CraftingContainer;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
@@ -19,7 +19,6 @@ import net.minecraft.world.level.Level;
 
 public class CopyBlockDyeRecipe extends CustomRecipe {
     public static final String REGISTRY_NAME = "copyblockdye";
-    public static final String COPY_TAG = AColoredBlock.DYE_ID.getName();
 
 	public static final SimpleCraftingRecipeSerializer<?> SERIALIZER = new SimpleCraftingRecipeSerializer<CopyBlockDyeRecipe>(CopyBlockDyeRecipe::new);
 
@@ -59,7 +58,7 @@ public class CopyBlockDyeRecipe extends CustomRecipe {
     }
 
     @Override
-    public ItemStack assemble(CraftingContainer inv, RegistryAccess r) {
+    public ItemStack assemble(CraftingContainer inv, HolderLookup.Provider pRegistries) {
         ItemStack source = ItemStack.EMPTY;
         ItemStack target = ItemStack.EMPTY;
 
@@ -74,10 +73,9 @@ public class CopyBlockDyeRecipe extends CustomRecipe {
         }
         
         if(!source.isEmpty() && !target.isEmpty()) {
-            CompoundTag targetTag = target.getOrCreateTag().copy();
-            targetTag.putInt(COPY_TAG, TagUtils.getIntOrDefault(source, COPY_TAG, 0));
             ItemStack result = target.copyWithCount(1);
-            result.setTag(targetTag);
+            DataComponentType<Integer> DYE_COMPONENT = TagUtils.getIntComponent(AColoredBlock.DYE_ID.getName());
+            result.set(DYE_COMPONENT, TagUtils.getOrDefault(source, DYE_COMPONENT, 0));
             return result;
         }
 
