@@ -5,6 +5,7 @@ import javax.annotation.Nullable;
 
 import io.github.tofodroid.mods.mimi.util.MidiNbtDataUtils;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.item.ItemStack;
@@ -20,7 +21,6 @@ public abstract class AConfigurableMidiNoteResponsiveTile extends AConfigurableM
     protected Byte filterNote;
     protected Byte filteredInstrument;
     protected Boolean invertFilterNoteOct;
-    protected Boolean invertFilterInstrument;
 
     public static void doTick(Level world, BlockPos pos, BlockState state, AConfigurableMidiNoteResponsiveTile self) {
         self.tick(world, pos, state);
@@ -47,8 +47,8 @@ public abstract class AConfigurableMidiNoteResponsiveTile extends AConfigurableM
     }
     
 	@Override
-	public void load(CompoundTag nbt) {
-        super.load(nbt);
+	public void loadAdditional(CompoundTag nbt, HolderLookup.Provider pRegistries) {
+        super.loadAdditional(nbt, pRegistries);
         this.cacheMidiSettings();
     }
 
@@ -93,7 +93,6 @@ public abstract class AConfigurableMidiNoteResponsiveTile extends AConfigurableM
         this.filterOctMax = (filterOct+1) * 12;
         this.filterNote = MidiNbtDataUtils.getFilterNote(this.getSourceStack());
         this.filteredInstrument = MidiNbtDataUtils.getFilterInstrument(this.getSourceStack());
-        this.invertFilterInstrument = MidiNbtDataUtils.getInvertInstrument(this.getSourceStack());
         this.invertFilterNoteOct = MidiNbtDataUtils.getInvertNoteOct(this.getSourceStack());
     }
 
