@@ -4,10 +4,14 @@ import java.util.ArrayList;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import io.github.tofodroid.mods.mimi.util.TagUtils;
 import io.github.tofodroid.mods.mimi.server.midi.playlist.APlaylistHandler.FavoriteMode;
 import io.github.tofodroid.mods.mimi.server.midi.playlist.APlaylistHandler.LoopMode;
 import io.github.tofodroid.mods.mimi.server.midi.playlist.APlaylistHandler.SourceMode;
+import net.minecraft.core.HolderLookup.Provider;
+import net.minecraft.core.component.DataComponentHolder;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.saveddata.SavedData;
 
 public class PlaylistData extends SavedData {
@@ -18,29 +22,24 @@ public class PlaylistData extends SavedData {
     public Boolean isShuffled = false;
     public Integer shuffleSeed = 0;
 
-    @Override
-    public CompoundTag save(CompoundTag tag) {
-        return PlaylistData.writeToTag(this, tag);
-    }
-
-    public static PlaylistData loadFromTag(CompoundTag tag) {
+    public static PlaylistData loadFromTag(CompoundTag tag, Provider registries) {
         PlaylistData data = new PlaylistData();
 
-        if(tag.contains("FAVORITE_SONGS")) {
-            CompoundTag favoriteSongTag = tag.getCompound("FAVORITE_SONGS");
+        if(tag.contains("favorite_songs")) {
+            CompoundTag favoriteSongTag = tag.getCompound("favorite_songs");
             data.favoriteSongs = new ArrayList<>(favoriteSongTag.getAllKeys().stream().map(key -> UUID.fromString(key)).collect(Collectors.toList()));
         }
 
-        if(tag.contains("LOOP_MODE")) {
-            data.loopMode = tag.getBoolean("LOOP_MODE") ? LoopMode.ALL : LoopMode.SINGLE;
+        if(tag.contains("loop_mode")) {
+            data.loopMode = tag.getBoolean("loop_mode") ? LoopMode.ALL : LoopMode.SINGLE;
         }
         
-        if(tag.contains("FAVORITE_MODE")) {
-            data.favoriteMode = tag.getBoolean("FAVORITE_MODE") ? FavoriteMode.FAVORITE : FavoriteMode.NOT_FAVORITE;
+        if(tag.contains("favorite_mode")) {
+            data.favoriteMode = tag.getBoolean("favorite_mode") ? FavoriteMode.FAVORITE : FavoriteMode.NOT_FAVORITE;
         }
         
-        if(tag.contains("SOURCE_MODE")) {
-            data.sourceMode = tag.getBoolean("SOURCE_MODE") ? SourceMode.CLIENT : SourceMode.SERVER;
+        if(tag.contains("source_mode")) {
+            data.sourceMode = tag.getBoolean("source_mode") ? SourceMode.CLIENT : SourceMode.SERVER;
         }
 
         if(tag.contains("SHUFFLED")) {
@@ -51,34 +50,34 @@ public class PlaylistData extends SavedData {
         return data;
     }
 
-    public static CompoundTag writeToTag(PlaylistData data, CompoundTag resultTag) {
+    public static CompoundTag writeToTag(PlaylistData data, CompoundTag resultTag, Provider registries) {
         CompoundTag favoriteSongTag = new CompoundTag();
 
         if(!data.favoriteSongs.isEmpty()) {
             for(Integer i = 0; i < data.favoriteSongs.size(); i++) {
                 favoriteSongTag.putBoolean(data.favoriteSongs.get(i).toString(), true);
             }
-            resultTag.put("FAVORITE_SONGS", favoriteSongTag);
+            resultTag.put("favorite_songs", favoriteSongTag);
         } else {
-            resultTag.remove("FAVORITE_SONGS");
+            resultTag.remove("favorite_songs");
         }
 
         if(data.loopMode != LoopMode.NONE) {
-            resultTag.putBoolean("LOOP_MODE", data.loopMode == LoopMode.ALL);
+            resultTag.putBoolean("loop_mode", data.loopMode == LoopMode.ALL);
         } else {
-            resultTag.remove("LOOP_MODE");
+            resultTag.remove("loop_mode");
         }
 
         if(data.favoriteMode != FavoriteMode.ALL) {
-            resultTag.putBoolean("FAVORITE_MODE", data.favoriteMode == FavoriteMode.FAVORITE);
+            resultTag.putBoolean("favorite_mode", data.favoriteMode == FavoriteMode.FAVORITE);
         } else {
-            resultTag.remove("FAVORITE_MODE");
+            resultTag.remove("favorite_mode");
         }
 
         if(data.sourceMode != SourceMode.ALL) {
-            resultTag.putBoolean("SOURCE_MODE", data.sourceMode == SourceMode.CLIENT);
+            resultTag.putBoolean("source_mode", data.sourceMode == SourceMode.CLIENT);
         } else {
-            resultTag.remove("SOURCE_MODE");
+            resultTag.remove("source_mode");
         }
 
         if(data.isShuffled) {
@@ -88,5 +87,59 @@ public class PlaylistData extends SavedData {
         }
 
         return resultTag;
+    }
+
+    public static PlaylistData loadFromComponents(DataComponentHolder components) {
+        PlaylistData data = new PlaylistData();
+
+        CompoundTag cval = TagUtils.getNbtOrDefault(components, "favorite_songs", null);
+        if(cval != null) {
+            data.favoriteSongs = new ArrayList<>(cval.getAllKeys().stream().map(key -> UUID.fromString(key)).collect(Collectors.toList()));
+        }
+
+        Boolean bval = TagUtils.getBooleanOrDefault(components, "loop_mode", null);
+        if(bval != null) {
+            data.loopMode = bval ? LoopMode.ALL : LoopMode.SINGLE;
+        }
+
+        bval = TagUtils.getBooleanOrDefault(components, "favorite_mode", null);
+        if(bval != null) {
+            data.favoriteMode = bval ? FavoriteMode.FAVORITE : FavoriteMode.FAVORITE;
+        }
+
+        bval = TagUtils.getBooleanOrDefault(components, "source_mode", null);
+        if(bval != null) {
+            data.sourceMode = bval ? SourceMode.CLIENT : SourceMode.SERVER;
+        }
+
+        Integer ival = TagUtils.getIntOrDefault(components, "shuffle", null);
+        if(ival != null) {
+            data.shuffleSeed = ival;
+        }
+
+        return data;
+    }
+
+    public static ItemStack writeToComponents(PlaylistData data, ItemStack stack) {
+        CompoundTag favoriteSongTag = null;
+        if(!data.favoriteSongs.isEmpty()) {
+            favoriteSongTag = new CompoundTag();
+            for(Integer i = 0; i < data.favoriteSongs.size(); i++) {
+                favoriteSongTag.putBoolean(data.favoriteSongs.get(i).toString(), true);
+            }
+        }
+        
+        TagUtils.setOrRemoveNbt(stack, "favorite_songs", favoriteSongTag);
+        TagUtils.setOrRemoveBoolean(stack, "loop_mode", data.loopMode == LoopMode.NONE ? null : data.loopMode == LoopMode.ALL);
+        TagUtils.setOrRemoveBoolean(stack, "favorite_mode", data.favoriteMode == FavoriteMode.ALL ? null : data.favoriteMode == FavoriteMode.FAVORITE);
+        TagUtils.setOrRemoveBoolean(stack, "source_mode", data.sourceMode == SourceMode.ALL ? null : data.sourceMode == SourceMode.CLIENT);
+        TagUtils.setOrRemoveInt(stack, "SHUFFLED", data.isShuffled ? data.shuffleSeed : null);
+
+        return stack;
+    }
+
+    @Override
+    public CompoundTag save(CompoundTag tag, Provider registries) {
+        return PlaylistData.writeToTag(this, tag, registries);
     }
 }
